Adding a constants file for colors and functions
