+++ src/components/SuggestedValidators/ValidatorTile.js
@@ -1,19 +1,11 @@
 import React from "react";
 import { Avatar, Box, Text, Heading, Flex, Badge } from "@chakra-ui/core";
-
-const borderColor = { light: "gray.100", dark: "#262E3F" };
-const textColor = { light: "gray.600", dark: "#7385A7" };
-const textColorLight = { light: "gray.300", dark: "#4B5871" };
-
-function getBadgeColor(risk) {
-	if (risk < 0.25) {
-		return "green";
-	}
-	if (risk < 0.5) {
-		return "yellow";
-	}
-	return "red";
-}
+import {
+	getRiskLevelColor as getBadgeColor,
+	textColor,
+	textColorLight,
+	border as borderColor
+} from "../../constants";
 
 type ValidatorTileProps = {
 	colorMode: string,
